realocando arquivos
Em razão de deixar o projeto mais organizado.
