+++ src/styles/App.css
@@ -1,23 +1,25 @@
-.Tabuleiro{
+.Tabuleiro {
   display: flex;
   justify-content: center;
   flex-wrap: wrap;
   width: 700px;
-  
 }
 
-.App{
+.App {
   display: flex;
   justify-content: center;
   align-items: center;
   flex-direction: column;
   height: 100vh;
   width: 100vw;
-  background-color: #1C8394;
-
+  background-color: #1c8394;
 }
 
-button{
+.App > h1 {
+  margin: 0px;
+}
+
+button {
   font-size: 1.5rem;
   padding: 1rem 2rem;
   border: none;
@@ -26,5 +28,3 @@ button{
   cursor: pointer;
   transition: all 0.2s ease-in-out;
 }
-
-
